fonts attribute: https://github.com/metanorma/metanorma-standoc/issues/618
